Add files via upload
The dataset contains information about symptoms of heart diseases. The data can be used for a lot of purposes such as disease prediction to exemplify the use of logistic regression in Machine Learning. The dataset was not in good format and contain a lot of hidden information. In order to train the model preprocessing on the data has done to refine it into required shape. After preprocessing and model training and choosing the right model with maximum accuracy a lot of models is train and finally the model with maximum accuracy was chosen to get accurate prediction.